+++ src/registry/index.js
@@ -1,0 +1,6 @@
+/** @flow */
+import npmLogin from './registry';
+
+export {
+  npmLogin,
+};